+++ android_client/app/src/main/java/com/snowtraces/noteall/model/AppView.kt
@@ -1,3 +1,3 @@
 package com.snowtraces.noteall.model
 
-enum class AppView { Home, Trash }
+enum class AppView { Home, Trash, Chat, ChatSessions }